code/requirements.txt
absl-py==1.4.0
asttokens==2.2.1
backcall==0.2.0
batchgenerators==0.25
cached-property==1.5.2
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.2.0
colorama==0.4.6
coloredlogs==15.0.1
contextlib2==21.6.0
cycler==0.11.0
decorator==5.1.1
einops==0.6.1
executing==1.2.0
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.38.0
fsspec==2023.1.0
future==0.18.3
h5py==3.5.0
huggingface-hub==0.16.4
humanfriendly==10.0
icecream==2.1.3
idna==3.4
imageio==2.10.1
importlib-metadata==6.7.0
ipdb==0.13.13
ipython==7.34.0
jedi==0.19.0
joblib==1.3.2
kiwisolver==1.4.4
linecache2==1.0.0
matplotlib==3.5.3
matplotlib-inline==0.1.6
MedPy==0.4.0
ml-collections==0.1.1
monai==1.1.0
mpmath==1.3.0
networkx==2.6.3
nibabel==4.0.2
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
onnx==1.13.1
onnxruntime==1.14.1
opencv-python==4.5.4.58
packaging==23.1
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.39
protobuf==3.20.3
ptyprocess==0.7.0
pycocotools==2.0.6
Pygments==2.15.1
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.3.0
PyYAML==6.0.1
requests==2.31.0
safetensors==0.3.1
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
SimpleITK==2.2.1
six==1.16.0
sympy==1.10.1
tensorboardX==2.6
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.9.5
tomli==2.0.1
torch==1.9.1+cu111
torchvision==0.10.1+cu111
tqdm==4.62.3
traceback2==1.4.0
traitlets==5.9.0
typing_extensions==4.7.1
unittest2==1.1.0
urllib3==2.0.4
wcwidth==0.2.6
yamlloader==1.3.2
zipp==3.15.0
